CC-5380 : Pypo should support subseconds for crossfades, general scheduling
removing useless/duplicate functions.
sending subseconds to pypo
standardizing changing everything from HH:mm:ss.u to U.u
starting to use only precision math for scheduling calculations

+++ airtime_mvc/application/models/ShowInstance.php
@@ -620,7 +620,7 @@ SQL;
     {
         $time_filled = $this->getTimeScheduled();
 
-        return Application_Common_DateHelper::playlistTimeToSeconds($time_filled);
+        return (float) Application_Common_DateHelper::playlistTimeToSeconds($time_filled);
     }
 
     public function getDurationSecs()
